--- .eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: ['@kyeotic/eslint-config'],
-  rules: {
-    'react/display-name': 0
-  }
-}
