+++ .github/workflows/test.yaml
@@ -1,40 +1,40 @@
 name: Tests
 
 on:
-  push:
-    branches:
-    - master
-    paths-ignore:
-      - 'docs/**'
-  pull_request:
-    branches:
-    - master
-    paths-ignore:
-      - 'docs/**'
+    push:
+        branches:
+            - master
+        paths-ignore:
+            - 'docs/**'
+    pull_request:
+        branches:
+            - master
+        paths-ignore:
+            - 'docs/**'
 
 jobs:
-  build:
-    runs-on: ubuntu-latest
+    build:
+        runs-on: ubuntu-latest
 
-    strategy:
-      matrix:
-        node-version: [10.x, 12.x, 14.x]
+        strategy:
+            matrix:
+                node-version: [10.x, 12.x, 14.x]
 
-    steps:
-      - uses: actions/checkout@v1
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v1
-        with:
-          node-version: ${{ matrix.node-version }}
-      - name: npm install, build, and test
-        run: |
-          npm install
-          npm run build --if-present
-          npm run test:ci
-        env:
-          CI: true
-      - name: Archive code coverage results
-        uses: actions/upload-artifact@v1
-        with:
-          name: code-coverage-report
-          path: junit.xml
+        steps:
+            - uses: actions/checkout@v1
+            - name: Use Node.js ${{ matrix.node-version }}
+              uses: actions/setup-node@v1
+              with:
+                  node-version: ${{ matrix.node-version }}
+            - name: yarn install, build, and test
+              run: |
+                  yarn install
+                  yarn run build --if-present
+                  yarn run test:ci
+              env:
+                  CI: true
+            - name: Archive code coverage results
+              uses: actions/upload-artifact@v1
+              with:
+                  name: code-coverage-report
+                  path: junit.xml
