+++ .gitignore
@@ -1,42 +1,6 @@
-lib-cov
-*.seed
 *.log
-*.csv
-*.dat
-*.out
-*.pid
-*.gz
-
-pids
-logs
-results
-
-npm-debug.log
-
-.DS_store
-
-*.sublime-workspace
-node_modules/
-main-built.js
-
-dist/*
-example/dist/*
-.history/
-/build
-/public/
-/themes/
-
-# misc
-.idea
 .DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.vscode/
-coverage/
-/junit.xml
+node_modules
+.cache
+.parcel-cache
+dist
