--- babel.config.js
@@ -1,15 +1,0 @@
-// This is only used for testing with babel-jest
-
-module.exports = {
-  presets: [
-    '@babel/preset-react',
-    [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current'
-        }
-      }
-    ]
-  ]
-}
