+++ example/index.html
@@ -1,18 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+        <title>Playground</title>
+    </head>
 
-<head>
-	<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-	<title>Raviger</title>
-	<script src="/dist/react.js"></script>
-	<script src="/dist/react-dom.js"></script>
-	<script src="/dist/whyDidYouRender.js"></script>
-	<link href="/src/app.css" rel="stylesheet"></link>
-</head>
-
-<body>
-	<div id="root"></div>
-	<script src="/dist/app.js"></script>
-</body>
-
+    <body>
+        <div id="root"></div>
+        <script src="./src/App.tsx"></script>
+    </body>
 </html>
