+++ example/package.json
@@ -1,0 +1,24 @@
+{
+    "name": "example",
+    "version": "1.0.0",
+    "main": "index.js",
+    "license": "MIT",
+    "scripts": {
+        "start": "parcel index.html",
+        "build": "parcel build index.html"
+    },
+    "dependencies": {
+        "react-app-polyfill": "^1.0.0"
+    },
+    "alias": {
+        "react": "../node_modules/react",
+        "react-dom": "../node_modules/react-dom/profiling",
+        "scheduler/tracing": "../node_modules/scheduler/tracing-profiling"
+    },
+    "devDependencies": {
+        "@types/react": "^16.9.11",
+        "@types/react-dom": "^16.8.4",
+        "parcel": "^2.0.0-beta.2",
+        "typescript": "^3.4.5"
+    }
+}
