+++ package.json
@@ -1,17 +1,7 @@
 {
   "name": "raviger",
-  "version": "2.0.2",
+  "version": "3.0.0",
   "description": "React routing with hooks",
-  "main": "dist/main.js",
-  "module": "dist/module.js",
-  "files": [
-    "dist/",
-    "types/"
-  ],
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "types": "types/raviger.d.ts",
   "keywords": [
     "react",
     "hooks",
@@ -25,77 +15,88 @@
   "author": "Tim Kye <tyrsius@gmail.com>",
   "license": "MIT",
   "private": false,
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
   "scripts": {
-    "style": "prettier --config package.json  --write \"{src,test}/**/*.js\"",
-    "style:ci": "prettier --config package.json --check \"{src,test}/**/*.js\"",
-    "lint": "eslint -c package.json \"{src,test}/**/*.js\"",
-    "check": "npm run style && npm run lint",
-    "check:ci": "npm run style:ci && npm run lint",
-    "build": "rollup -c",
-    "build:watch": "rollup -c --watch",
+    "style": "prettier --config package.json  --write \"{src,test}/**/*.{ts,tsx}\"",
+    "style:ci": "prettier --config package.json --check \"{src,test}/**/*.{ts,tsx}\"",
+    "check": "yarn run style && yarn run lint",
+    "check:ci": "yarn run style:ci && yarn run lint",
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "build:watch": "yarn run start",
     "example:build": "NODE_ENV=development rollup -c example/rollup.config.js && yarn run example:build:link-react",
     "example:build:watch": "rollup -c example/rollup.config.js --watch",
     "example:build:link-react": "mkdir -p example/dist && ln -sf $PWD/node_modules/react/umd/react.development.js example/dist/react.js && ln -sf $PWD/node_modules/react-dom/umd/react-dom.development.js example/dist/react-dom.js && ln -sf $PWD/node_modules/@welldone-software/why-did-you-render/dist/umd/whyDidYouRender.js example/dist/whyDidYouRender.js",
     "example:serve": "npm run example:build:link-react && http-server example/ -p 8080 --proxy http://localhost:8080?",
     "example:dev": "run-p example:build:watch example:serve",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
+    "test": "tsdx test --passWithNoTests",
+    "test:watch": "tsdx test --watch",
+    "test:coverage": "tsdx test --coverage",
     "test:coverage:open": "run-s test:coverage coverage:report",
-    "test:unit:ci": "jest --ci --coverage --reporters=default --reporters=jest-junit",
+    "test:unit:ci": "tsdx test --ci --coverage --reporters=default --reporters=jest-junit",
     "coverage:report": "open coverage/lcov-report/index.html",
     "test:ci": "run-s check:ci test:unit:ci size",
     "size": "run-s build size:check",
     "size:check": "size-limit",
     "prepublishOnly": "run-s test:ci",
-    "release": "np"
+    "release": "np",
+    "lint": "tsdx lint",
+    "prepare": "tsdx build",
+    "analyze": "size-limit --why"
   },
   "peerDependencies": {
-    "react": "^16.8.0"
+    "react": ">=16"
   },
-  "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/preset-env": "^7.8.4",
-    "@babel/preset-react": "^7.8.3",
-    "@kyeotic/eslint-config": "^1.0.2",
-    "@size-limit/preset-small-lib": "^4.0.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^8.0.7",
-    "@types/jest": "^24.9.1",
-    "@welldone-software/why-did-you-render": "^3.6.0",
-    "babel-jest": "^24.8.0",
-    "core-js": "^3.6.5",
-    "eslint": "^5.6.0",
-    "eslint-plugin-jest": "^22.21.0",
-    "eslint-plugin-react": "^7.18.3",
-    "http-server": "^0.11.1",
-    "jest": "^24.8.0",
-    "jest-junit": "^8.0.0",
-    "np": "^6.2.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^1.19.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "rollup": "^1.31.0",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "size-limit": "^4.0.0"
-  },
-  "np": {
-    "yarn": false
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
   },
   "prettier": {
-    "tabWidth": 2,
+    "tabWidth": 4,
+    "useTabs": false,
+    "trailingComma": "all",
+    "bracketSpacing": false,
     "semi": false,
     "singleQuote": true,
     "printWidth": 80
   },
+  "module": "dist/raviger.esm.js",
   "size-limit": [
     {
+      "path": "dist/raviger.cjs.production.min.js",
+      "limit": "2.5kb"
+    },
+    {
       "webpack": false,
-      "path": "dist/module.js",
+      "path": "dist/raviger.esm.js",
       "limit": "2.5kb"
     }
-  ]
+  ],
+  "np": {
+    "yarn": false
+  },
+  "devDependencies": {
+    "@size-limit/preset-small-lib": "^4.10.2",
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/react": "^11.2.6",
+    "@types/react": "^17.0.4",
+    "@types/react-dom": "^17.0.3",
+    "core-js": "^3.11.0",
+    "husky": "^6.0.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "size-limit": "^4.10.2",
+    "tsdx": "^0.14.1",
+    "tslib": "^2.2.0",
+    "typescript": "^4.2.4"
+  }
 }
