--- src/main.js
@@ -1,12 +1,0 @@
-export { useRoutes } from './router.js'
-export { useRedirect, Redirect } from './redirect.js'
-export { default as Link, ActiveLink } from './Link.js'
-export { navigate, useNavigate, useNavigationPrompt } from './navigate.js'
-export {
-  usePath,
-  useHash,
-  useFullPath,
-  useBasePath,
-  useLocationChange
-} from './path.js'
-export { useQueryParams } from './querystring.js'
