+++ src/node.ts
@@ -1,13 +1,13 @@
 let ssrPath = '/'
 let isNode = true // eslint-disable-line import/no-mutable-exports
 try {
-  isNode = window === undefined
+    isNode = window === undefined
 } catch (e) {} // eslint-disable-line no-empty
 
-export { isNode }
+export {isNode}
 export function getSsrPath() {
-  return ssrPath
+    return ssrPath
 }
-export function setSsrPath(path) {
-  ssrPath = path
+export function setSsrPath(path: string) {
+    ssrPath = path
 }
