--- src/typeChecks.js
@@ -1,3 +1,0 @@
-export function isFunction(obj) {
-  return obj && typeof obj === 'function'
-}
